IndustrialManagement_Backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.29
psycopg2==2.9.10
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
